+++ crates/cli-tools/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Minor
 
+- Change the flags of `action::AppletRpc` to use `action::Wait`
+- Add `action::Wait` for commands returning an optional response
 - Increase the default connection timeout from 1 to 5 seconds
 - Add `action::PlatformUpdate` for platform update
 - Add `action::PlatformList` to list connected platforms